general FunctionChart refactorings
